backend_django/requirements.txt
Django==4.2.14
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.2.0
celery==5.3.6
redis==4.6.0
django-celery-results==2.5.1
django-celery-beat==2.8.0
gunicorn==21.2.0
channels>=4.0.0
djangorestframework-simplejwt==5.3.1
